requirements.txt
pygame==2.5.0
moderngl==5.8.2
numba==0.57.1
PyGLM==2.7.0
numpy<1.25,>=1.21
opensimplex[numpy]==0.4.4